Learning the concept of rect transform
The first to making the project is creating the UI based on the design that has been decided. In order to deepen my understanding on the rect transform and how can I manage it well, I tried learning how does a rect transform works. As a result, here is the UI, although it is not clean, I am confident that the hierarchy is well organized and scalable